Fixing generic ssh_keys_auth module
